Moved spectrum to its own package
Currently the opacity state fixture needs to be moved
