+++ benchmarks/transport_montecarlo_numba_formal_integral_p.py
@@ -6,7 +6,7 @@ import numpy as np
 from asv_runner.benchmarks.mark import parameterize
 from numba import config
 
-import tardis.transport.montecarlo.formal_integral as formal_integral
+import tardis.spectrum.formal_integral as formal_integral
 from benchmarks.benchmark_base import BenchmarkBase
 from tardis import constants as c
 from tardis.model.geometry.radial1d import NumbaRadial1DGeometry
