+++ tardis/spectrum/tests/test_numba_formal_integral.py
@@ -6,7 +6,7 @@ from copy import deepcopy
 import numpy.testing as ntest
 
 from tardis.util.base import intensity_black_body
-import tardis.transport.montecarlo.formal_integral as formal_integral
+import tardis.spectrum.formal_integral as formal_integral
 from tardis.model.geometry.radial1d import NumbaRadial1DGeometry
