+++ tardis/transport/montecarlo/base.py
@@ -19,6 +19,10 @@ from tardis.transport.montecarlo.estimators.radfield_mc_estimators import (
     initialize_estimator_statistics,
 )
 from tardis.transport.montecarlo.formal_integral import FormalIntegrator
+from tardis.transport.montecarlo.montecarlo_configuration import (
+    MonteCarloConfiguration,
+    configuration_initialize,
+)
 from tardis.transport.montecarlo.montecarlo_transport_state import (
     MonteCarloTransportState,
 )
@@ -187,9 +191,9 @@ class MonteCarloTransportSolver(HDFWriterMixin):
             total_iterations=total_iterations,
         )
 
-        transport_state._montecarlo_virtual_luminosity.value[
-            :
-        ] = v_packets_energy_hist
+        transport_state._montecarlo_virtual_luminosity.value[:] = (
+            v_packets_energy_hist
+        )
         transport_state.last_interaction_type = last_interaction_tracker.types
         transport_state.last_interaction_in_nu = last_interaction_tracker.in_nus
         transport_state.last_interaction_in_r = last_interaction_tracker.in_rs
